Extract useKeyboardShortcuts composable from EditorApp.vue
Move keyboard shortcut handling (~120 lines) into dedicated composable:
- handleKeyDown with all shortcut logic (slot switch, undo/redo, copy/paste/cut, delete, escape)
- onMounted/onUnmounted lifecycle for event listener registration
- Accepts action callbacks via options object for loose coupling

Zero functional changes - pure refactoring.
